moved manual_control into the miniworld package
